Remove unneeded target frameworks
If you target .NET Framework 4.5, you also allow that
package to work on versions of .NET Framework 4.x
where x is larger than 5. And so is the case
with .NET Core 2.0. .NET Standard 2.0 is, OTOH,
separate from .NET Standard 1.6, because the latter
adds their dependency on the NuGet packages.

The only thing that is accomplished by adding
these is that the user compiling the library has
to install more targeting packs on their machine.